wine user ownership finished

+++ src/Controller/WineController.php
@@ -22,6 +22,14 @@ class WineController extends AbstractController
     {
         $this->security = $security;
     }
+    public function getSessionUserId()
+    {
+        $userId = '';
+        if ($this->security->getUser()) {
+            $userId = $this->security->getUser()->getId();
+        }
+        return $userId;
+    }
 
     /**
      * @Route("/new", name="wine_new", methods={"GET","POST"})
@@ -29,12 +37,11 @@ class WineController extends AbstractController
     public function new(Request $request, UploaderHelper $uploaderHelper): Response
     {
         $wine = new Wine();
-        $user = $this->security->getUser();
         $form = $this->createForm(WineType::class, $wine);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $wine->setAuthor($user->getId());
+            $wine->setAuthor($this->getSessionUserId());
             $uploadedFile = $form['imageFilename']->getData(); 
             if ($uploadedFile) {
                 $newFilename = $uploaderHelper->uploadArticleImage($uploadedFile);
@@ -63,25 +70,26 @@ class WineController extends AbstractController
         $form = $this->createForm(WineType::class, $wine);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            
-            $currentFilename = $wine->getImageFilename();
-            $uploadedFile = $form['imageFilename']->getData(); 
-            if ($uploadedFile) {
-                $newFilename = $uploaderHelper->uploadArticleImage($uploadedFile);
+            if ($form->isSubmitted() && $form->isValid() && $this->getSessionUserId() == $wine->getAuthor()) {
+                $currentFilename = $wine->getImageFilename();
+                $uploadedFile = $form['imageFilename']->getData(); 
+                if ($uploadedFile) {
+                    $newFilename = $uploaderHelper->uploadArticleImage($uploadedFile);
 
-                if ($newFilename) {
-                    $wine->setImageFilename($newFilename);
-                    $uploaderHelper->removeFile(UploaderHelper::IMAGES, $currentFilename);
+                    if ($newFilename) {
+                        $wine->setImageFilename($newFilename);
+                        $uploaderHelper->removeFile(UploaderHelper::IMAGES, $currentFilename);
+                    }
                 }
-            }
-            $this->getDoctrine()->getManager()->flush();
+                $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('wine');
-        }
+                return $this->redirectToRoute('wine');
+            }
+
         return $this->render('wine/edit.html.twig', [
             'wine' => $wine,
             'form' => $form->createView(),
+            'userId' => $this->getSessionUserId(),
         ]);
     }
 
@@ -90,7 +98,7 @@ class WineController extends AbstractController
      */ 
     public function delete(Request $request, Wine $wine, UploaderHelper $uploaderHelper): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$wine->getId(), $request->request->get('_token'))) {
+        if ($this->isCsrfTokenValid('delete'.$wine->getId(), $request->request->get('_token')) && $this->getSessionUserId() == $wine->getAuthor()) {
             $uploaderHelper->removeFile(UploaderHelper::IMAGES, $wine->getImageFilename());
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($wine);
